src/requirements.txt
OpenChemIE @ git+https://github.com/CrystalEye42/OpenChemIE
requests
numpy
opencv-python-headless
Pillow